+++ Code/Editor/EditorFileMonitor.cpp
@@ -143,8 +143,8 @@ QString RemoveGameName(const QString &filename)
     // Remove first part of path.  File coming in has the game name included
     // eg (AutomatedTesting/Animations/Chicken/anim_chicken_flapping.i_caf)->(Animations/Chicken/anim_chicken_flapping.i_caf)
 
-    int indexOfFirstSlash = filename.indexOf('/');
-    int indexOfFirstBackSlash = filename.indexOf('\\');
+    int indexOfFirstSlash = static_cast<int>(filename.indexOf('/'));
+    int indexOfFirstBackSlash = static_cast<int>(filename.indexOf('\\'));
     if (indexOfFirstSlash >= 0)
     {
         if (indexOfFirstBackSlash >= 0 && indexOfFirstBackSlash < indexOfFirstSlash)
